shill: Fix type for tag in RoutingTableEntry
|tag| should be int since it will be set to the value of an interface
index. Currently there will be some issues with routing if there are
(used to be) too many interfaces on the DUT that causes the interface
index exceeding 256.

BUG=None
TEST=unit tests

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3702821
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>

+++ shill/routing_table_entry.cc
@@ -59,7 +59,7 @@ RoutingTableEntry& RoutingTableEntry::SetType(unsigned char type_in) {
   return *this;
 }
 
-RoutingTableEntry& RoutingTableEntry::SetTag(uint8_t tag_in) {
+RoutingTableEntry& RoutingTableEntry::SetTag(int tag_in) {
   tag = tag_in;
   return *this;
 }
